dashboard/requirements.txt
kafka-python
rich